Add include folder
