+++ include/HfCloud.h
@@ -1,0 +1,12 @@
+#ifndef _HFCLOUD_H
+#define _HFCLOUD_H
+
+#include "stdinc.h"
+#include "Graphics.h"
+#include "Scene.h"
+#include "Sprite.h"
+#include "Input.h"
+#include "SceneManager.h"
+#include "Bitmap.h"
+
+#endif // _HFCLOUD_H
