+++ include/stdinc.h
@@ -1,0 +1,29 @@
+#ifndef _STD_INC_FILES
+#define _STD_INC_FILES
+
+// includes some headers.
+
+#include <cstdio>
+#include <cstring>
+#include <cstdlib>
+#include <ctime>
+#include <cmath>
+#include <algorithm>
+#include <set>
+#include <list>
+#include <queue>
+#include <deque>
+#include <vector>
+#include <iostream>
+#include <SDL.h>
+#include <SDL_image.h>
+typedef SDL_Rect Hf_Rect;
+typedef SDL_Color Hf_Color;
+typedef SDL_Point Hf_Point;
+
+namespace HfCloud{}
+
+#define HFCLOUD_BEGIN namespace HfCloud{
+#define HFCLOUD_END }
+
+#endif // _STD_INC_FILES
