+++ main.cpp
@@ -1,7 +1,7 @@
-#include "stdinc.h"
-#include "SceneManager.h"
-#include "Graphics.h"
-#include "Fiber.h"
+#include "include/stdinc.h"
+#include "include/SceneManager.h"
+#include "include/Graphics.h"
+#include "include/Fiber.h"
 
 using namespace HfCloud;
 const int WINDOW_WIDTH = 544; /**<The width of the game window*/
