+++ stdinc.cpp
@@ -1,4 +1,4 @@
-#include "stdinc.h"
+#include "include/stdinc.h"
 
 #ifdef _MSC_VER
